A 32bit cleanup for MDEV-14452 Precision in INTERVAL xxx DAY_MICROSECOND parsed wrong?
"mtr func_date_add" failed on 32-bit platforms. Removing a wrong case to "long".
Both values[] and log_10_int[] are arrays of "ulonglong", no cast is needed.

+++ sql/item_timefunc.cc
@@ -732,9 +732,9 @@ static bool get_interval_info(const char *str,uint length,CHARSET_INFO *cs,
   if (transform_msec && field_length > 0)
   {
     if (field_length < 6)
-      values[count - 1] *= (long) log_10_int[6 - field_length];
+      values[count - 1] *= log_10_int[6 - field_length];
     else if (field_length > 6)
-      values[count - 1] /= (long) log_10_int[field_length - 6];
+      values[count - 1] /= log_10_int[field_length - 6];
   }
 
   return (str != end);
